requirements.txt
python-dateutil>=2.8.1
reportportal-client>=5.2.3
robotframework
six>=1.15.0
